Update index.html
Check the number of items each time you try to add a new item, the query selector result doesn't update as the page changes. You probably want to count the number of li elements, not the number of inputs. Or do you mean `.input` with a dot - selecting by the class input?